requirements.txt
fastapi==0.123.5
uvicorn[standard]==0.38.0
httpx==0.28.1
python-dotenv==1.2.1
pydantic==2.12.5
